+++ src/main/java/org/vetmeduni/io/writers/fastq/ReadToolsFastqWriterFactory.java
@@ -29,8 +29,8 @@ import htsjdk.samtools.util.Lazy;
 import htsjdk.samtools.util.Log;
 import org.vetmeduni.io.FastqPairedRecord;
 import org.vetmeduni.io.IOdefault;
-import org.vetmeduni.methods.barcodes.dictionary.decoder.BarcodeDecoder;
 import org.vetmeduni.methods.barcodes.dictionary.BarcodeDictionary;
+import org.vetmeduni.methods.barcodes.dictionary.decoder.BarcodeMatch;
 import org.vetmeduni.utils.misc.IOUtils;
 import org.vetmeduni.utils.record.FastqRecordUtils;
 
@@ -170,8 +170,9 @@ public class ReadToolsFastqWriterFactory {
 			mapping.put(dictionary.getCombinedBarcodesFor(i), sampleNames.get(sample));
 		}
 		// add a unknow barcode
-		mapping.put(BarcodeDecoder.UNKNOWN_STRING,
-			(paired) ? newPairWriter(prefix + "_" + IOdefault.DISCARDED_SUFFIX) : newWriter(prefix + "_" + IOdefault.DISCARDED_SUFFIX));
+		mapping.put(BarcodeMatch.UNKNOWN_STRING, (paired) ?
+			newPairWriter(prefix + "_" + IOdefault.DISCARDED_SUFFIX) :
+			newWriter(prefix + "_" + IOdefault.DISCARDED_SUFFIX));
 		return new SplitFastqWriterAbstract(mapping) {
 
 			private final boolean p = paired;
@@ -207,7 +208,7 @@ public class ReadToolsFastqWriterFactory {
 			 * @return the barcode if it is present, the UNKNOWN_STRING otherwise
 			 */
 			private String getUnknownIfNoMapping(String barcode) {
-				return (mapping.contains(barcode)) ? barcode : BarcodeDecoder.UNKNOWN_STRING;
+				return (mapping.contains(barcode)) ? barcode : BarcodeMatch.UNKNOWN_STRING;
 			}
 		};
 	}
@@ -240,8 +241,8 @@ public class ReadToolsFastqWriterFactory {
 
 	/**
 	 * Writer that split between assign/unknow barcodes; the mapping is "assign" and {@link
-	 * org.vetmeduni.methods.barcodes.dictionary.decoder.BarcodeDecoder#UNKNOWN_STRING}. By default, any record is
-	 * correct unless the unknow string is provided as identifier
+	 * org.vetmeduni.methods.barcodes.dictionary.decoder.BarcodeMatch#UNKNOWN_STRING}. By default, any record is correct
+	 * unless the unknow string is provided as identifier
 	 *
 	 * @param prefix the prefix for the files
 	 * @param paired <code>true</code> indicates that paired writers are used; otherwise a default is used
@@ -252,8 +253,9 @@ public class ReadToolsFastqWriterFactory {
 		logger.debug("Creating new Assing-Unknown barcode for ", (paired) ? "paired" : "single", "-end");
 		final Hashtable<String, FastqWriter> mapping = new Hashtable<>(2);
 		mapping.put("assign", (paired) ? newPairWriter(prefix) : newWriter(prefix));
-		mapping.put(BarcodeDecoder.UNKNOWN_STRING,
-			(paired) ? newPairWriter(prefix + "_" + IOdefault.DISCARDED_SUFFIX) : newWriter(prefix + "_" + IOdefault.DISCARDED_SUFFIX));
+		mapping.put(BarcodeMatch.UNKNOWN_STRING, (paired) ?
+			newPairWriter(prefix + "_" + IOdefault.DISCARDED_SUFFIX) :
+			newWriter(prefix + "_" + IOdefault.DISCARDED_SUFFIX));
 		return new SplitFastqWriterAbstract(mapping) {
 
 			private final boolean p = paired;
@@ -293,7 +295,7 @@ public class ReadToolsFastqWriterFactory {
 
 			@Override
 			public void write(String identifier, FastqRecord record) {
-				if (BarcodeDecoder.UNKNOWN_STRING.equals(identifier)) {
+				if (BarcodeMatch.UNKNOWN_STRING.equals(identifier)) {
 					super.write(identifier, record);
 				} else {
 					write(record);
@@ -302,7 +304,7 @@ public class ReadToolsFastqWriterFactory {
 
 			@Override
 			public void write(String identifier, FastqPairedRecord record) {
-				if (BarcodeDecoder.UNKNOWN_STRING.equals(identifier)) {
+				if (BarcodeMatch.UNKNOWN_STRING.equals(identifier)) {
 					super.write(identifier, record);
 				} else {
 					write(record);