+++ src/main/java/org/vetmeduni/methods/barcodes/dictionary/BarcodeDictionaryFactory.java
@@ -23,6 +23,8 @@
 package org.vetmeduni.methods.barcodes.dictionary;
 
 import com.opencsv.CSVReader;
+import htsjdk.samtools.SAMReadGroupRecord;
+import org.vetmeduni.methods.barcodes.dictionary.decoder.BarcodeMatch;
 
 import java.io.File;
 import java.io.FileReader;
@@ -38,16 +40,32 @@ import java.util.Arrays;
 public class BarcodeDictionaryFactory {
 
 	/**
+	 * The read group information for default tags
+	 */
+	private static final SAMReadGroupRecord UNKNOWN_READGROUP_INFO;
+
+	// initialize the unknown read group information
+	static {
+		UNKNOWN_READGROUP_INFO = new SAMReadGroupRecord(BarcodeMatch.UNKNOWN_STRING);
+		UNKNOWN_READGROUP_INFO.setLibrary(BarcodeMatch.UNKNOWN_STRING);
+		UNKNOWN_READGROUP_INFO.setPlatform(BarcodeMatch.UNKNOWN_STRING);
+		UNKNOWN_READGROUP_INFO.setPlatformUnit(BarcodeMatch.UNKNOWN_STRING);
+		UNKNOWN_READGROUP_INFO.setSample(BarcodeMatch.UNKNOWN_STRING);
+	}
+
+	/**
 	 * Create a barcode dictionary from a tab-delimited file, with the first column being the barcode and the subsequent
 	 * the barcodes. Each of the barcodes is stored independently
 	 *
 	 * @param barcodeFile      the file
+	 * @param readGroupInfo    read group record where tags for other barcodes will be used (except ID and SN)
 	 * @param numberOfBarcodes the expected number of barcodes; if < 0, it is computed for the first line in the file
 	 *
 	 * @return the barcode dictionary
 	 * @throws java.io.IOException if the file have some problem
 	 */
-	public static BarcodeDictionary createDefaultDictionary(File barcodeFile, int numberOfBarcodes) throws IOException {
+	public static BarcodeDictionary createDefaultDictionary(File barcodeFile, final SAMReadGroupRecord readGroupInfo,
+		int numberOfBarcodes) throws IOException {
 		CSVReader reader = new CSVReader(new FileReader(barcodeFile), '\t');
 		// read the first line
 		String[] nextLine = reader.readNext();
@@ -67,7 +85,7 @@ public class BarcodeDictionaryFactory {
 		for (int i = 0; i < numberOfBarcodes; i++) {
 			barcodes.add(new ArrayList<>());
 		}
-		ArrayList<String> samples = new ArrayList<>();
+		ArrayList<SAMReadGroupRecord> samples = new ArrayList<>();
 		// reading the rest of the lines
 		while (nextLine != null) {
 			if (numberOfBarcodes != nextLine.length - 1) {
@@ -76,11 +94,69 @@ public class BarcodeDictionaryFactory {
 						.getAbsolutePath());
 			}
 			// the first item is the sample name
-			samples.add(nextLine[0]);
+			samples.add(getReadGroupForSample(nextLine[0], readGroupInfo));
 			// get the barcodes
 			for (int i = 1; i <= numberOfBarcodes; i++) {
 				barcodes.get(i - 1).add(nextLine[i]);
 			}
+			nextLine = reader.readNext();
+		}
+		reader.close();
+		// construct the barcode dictionary
+		return new BarcodeDictionary(samples, barcodes);
+	}
+
+	/**
+	 * Create a barcode dictionary from a tab-delimited file, with the first column being the barcode and the subsequent
+	 * the barcodes. Each of the barcodes is stored independently
+	 *
+	 * @param barcodeFile      the file
+	 * @param numberOfBarcodes the expected number of barcodes; if < 0, it is computed for the first line in the file
+	 *
+	 * @return the barcode dictionary
+	 * @throws java.io.IOException if the file have some problem
+	 */
+	public static BarcodeDictionary createDefaultDictionary(File barcodeFile, int numberOfBarcodes) throws IOException {
+		return createDefaultDictionary(barcodeFile, UNKNOWN_READGROUP_INFO, numberOfBarcodes);
+	}
+
+	/**
+	 * Creates a barcode dictionary from a tab-delimited file, with the first column being the barcode and the
+	 * subsequent the barcodes. All the barcodes are stored together as an "unique" barcode.
+	 *
+	 * @param barcodeFile   the file
+	 * @param readGroupInfo read group record where tags for other barcodes will be used (except ID and SN)
+	 *
+	 * @return the barcode dictionary
+	 * @throws IOException
+	 */
+	public static BarcodeDictionary createCombinedDictionary(File barcodeFile, final SAMReadGroupRecord readGroupInfo)
+		throws IOException {
+		CSVReader reader = new CSVReader(new FileReader(barcodeFile), '\t');
+		// read the first line
+		String[] nextLine = reader.readNext();
+		if (nextLine == null) {
+			throw new IOException("No data in the barcode file");
+		}
+		// compute the number of barcodes in the file
+		int numberOfBarcodes = nextLine.length - 1;
+		// the number of barcodes in the dictionary will be one
+		ArrayList<ArrayList<String>> barcodes = new ArrayList<>(1);
+		barcodes.add(new ArrayList<>());
+		// create the sample array
+		ArrayList<SAMReadGroupRecord> samples = new ArrayList<>();
+		// reading the rest of the lines
+		while (nextLine != null) {
+			if (numberOfBarcodes != nextLine.length - 1) {
+				throw new IOException(
+					"Each sample should have the same number of barcodes. Check the formatting of " + barcodeFile
+						.getAbsolutePath());
+			}
+			// the first item is the sample name
+			samples.add(getReadGroupForSample(nextLine[0], readGroupInfo));
+			// get the unique barcode
+			String uniqueBarcode = String.join("", Arrays.copyOfRange(nextLine, 1, nextLine.length));
+			barcodes.get(0).add(uniqueBarcode);
 			nextLine = reader.readNext();
 		}
 		reader.close();
@@ -98,35 +174,12 @@ public class BarcodeDictionaryFactory {
 	 * @throws IOException
 	 */
 	public static BarcodeDictionary createCombinedDictionary(File barcodeFile) throws IOException {
-		CSVReader reader = new CSVReader(new FileReader(barcodeFile), '\t');
-		// read the first line
-		String[] nextLine = reader.readNext();
-		if (nextLine == null) {
-			throw new IOException("No data in the barcode file");
-		}
-		// compute the number of barcodes in the file
-		int numberOfBarcodes = nextLine.length - 1;
-		// the number of barcodes in the dictionary will be one
-		ArrayList<ArrayList<String>> barcodes = new ArrayList<>(1);
-		barcodes.add(new ArrayList<>());
-		// create the sample array
-		ArrayList<String> samples = new ArrayList<>();
-		// reading the rest of the lines
-		while (nextLine != null) {
-			if (numberOfBarcodes != nextLine.length - 1) {
-				throw new IOException(
-					"Each sample should have the same number of barcodes. Check the formatting of " + barcodeFile
-						.getAbsolutePath());
-			}
-			// the first item is the sample name
-			samples.add(nextLine[0]);
-			// get the unique barcode
-			String uniqueBarcode = String.join("", Arrays.copyOfRange(nextLine, 1, nextLine.length));
-			barcodes.get(0).add(uniqueBarcode);
-			nextLine = reader.readNext();
-		}
-		reader.close();
-		// construct the barcode dictionary
-		return new BarcodeDictionary(samples, barcodes);
+		return createCombinedDictionary(barcodeFile, UNKNOWN_READGROUP_INFO);
+	}
+
+	private static SAMReadGroupRecord getReadGroupForSample(String sampleName, SAMReadGroupRecord info) {
+		SAMReadGroupRecord readGroup = new SAMReadGroupRecord(sampleName, info);
+		readGroup.setSample(sampleName);
+		return readGroup;
 	}
 }