+++ src/main/java/org/vetmeduni/tools/implemented/TaggedBamToFastq.java
@@ -31,17 +31,20 @@ import org.apache.commons.cli.Option;
 import org.apache.commons.cli.Options;
 import org.vetmeduni.io.FastqPairedRecord;
 import org.vetmeduni.io.writers.fastq.SplitFastqWriter;
-import org.vetmeduni.methods.barcodes.dictionary.decoder.BarcodeDecoder;
 import org.vetmeduni.methods.barcodes.dictionary.BarcodeDictionary;
 import org.vetmeduni.methods.barcodes.dictionary.BarcodeDictionaryFactory;
+import org.vetmeduni.methods.barcodes.dictionary.decoder.BarcodeDecoder;
+import org.vetmeduni.methods.barcodes.dictionary.decoder.BarcodeMatch;
 import org.vetmeduni.tools.AbstractTool;
 import org.vetmeduni.tools.cmd.CommonOptions;
 import org.vetmeduni.tools.cmd.ToolWritersFactory;
 import org.vetmeduni.tools.cmd.ToolsReadersFactory;
 import org.vetmeduni.utils.loggers.ProgressLoggerExtension;
+import org.vetmeduni.utils.misc.IOUtils;
 import org.vetmeduni.utils.record.SAMRecordUtils;
 
 import java.io.File;
+import java.io.IOException;
 
 import static org.vetmeduni.tools.ToolNames.ToolException;
 
@@ -66,7 +69,7 @@ public class TaggedBamToFastq extends AbstractTool {
 		String inputString = getUniqueValue(cmd, "i");
 		String outputPrefix = getUniqueValue(cmd, "o");
 		String barcodes = getUniqueValue(cmd, "bc");
-		int[] max = getIntArrayOptions(cmd.getOptionValues("m"), BarcodeDecoder.DEFAULT_MISMATCHES);
+		int[] max = getIntArrayOptions(cmd.getOptionValues("m"), BarcodeDecoder.DEFAULT_MAXIMUM_MISMATCHES);
 		String[] tags = cmd.getOptionValues("t");
 		logger.debug("Maximum mistmaches (", max.length, "): ", max);
 		logger.debug("Tags (", tags.length, "): ", tags);
@@ -82,7 +85,7 @@ public class TaggedBamToFastq extends AbstractTool {
 			.createDefaultDictionary(new File(barcodes), tags.length);
 		logger.info("Loaded barcode file for ", barcodeDict.numberOfUniqueSamples(), " samples with ",
 			barcodeDict.numberOfSamples(), " different barcode sets");
-		BarcodeDecoder matcher = new BarcodeDecoder(barcodeDict);
+		BarcodeDecoder decoder = new BarcodeDecoder(barcodeDict, max);
 		// open the bam file
 		SamReader input = ToolsReadersFactory
 			.getSamReaderFromInput(new File(inputString), CommonOptions.isMaintained(logger, cmd));
@@ -90,8 +93,10 @@ public class TaggedBamToFastq extends AbstractTool {
 		SplitFastqWriter writer = ToolWritersFactory
 			.getFastqSplitWritersFromInput(outputPrefix, cmd.hasOption("x") ? barcodeDict : null,
 				cmd.hasOption(CommonOptions.disableZippedOutput.getOpt()), multi, cmd.hasOption("s"));
+		// create the metrics file
+		File metrics = IOUtils.makeMetricsFile(outputPrefix);
 		// run it!
-		run(input, writer, matcher, max, tags, cmd.hasOption("s"));
+		run(input, writer, metrics, decoder, tags, cmd.hasOption("s"));
 		// close the readers and writers
 		input.close();
 		writer.close();
@@ -102,24 +107,24 @@ public class TaggedBamToFastq extends AbstractTool {
 	 *
 	 * @param reader  the input reader
 	 * @param writer  the output
-	 * @param matcher the matcher to use to split
-	 * @param max     the maximum number of mismatches
+	 * @param decoder the decoder to use to split
 	 * @param tags    the tags where the barcodes are
 	 * @param single  it is single end?
 	 */
-	private void run(SamReader reader, SplitFastqWriter writer, BarcodeDecoder matcher, int[] max,
-		String[] tags, boolean single) {
+	private void run(SamReader reader, SplitFastqWriter writer, File metrics, BarcodeDecoder decoder, String[] tags,
+		boolean single) throws IOException {
 		ProgressLoggerExtension progress;
 		// single end processing
 		if (single) {
 			progress = new ProgressLoggerExtension(logger, 1000000, "Processed", "records");
-			runSingle(reader, writer, matcher, max, tags, progress);
+			runSingle(reader, writer, decoder, tags, progress);
 		} else {
 			progress = new ProgressLoggerExtension(logger, 1000000, "Processed", "pairs");
-			runPaired(reader, writer, matcher, max, tags, progress);
+			runPaired(reader, writer, decoder, tags, progress);
 		}
 		progress.logNumberOfVariantsProcessed();
-		matcher.logMatcherResult(logger);
+		decoder.logMatcherResult(logger);
+		decoder.outputStats(metrics);
 	}
 
 	/**
@@ -128,11 +133,10 @@ public class TaggedBamToFastq extends AbstractTool {
 	 * @param reader  the input reader
 	 * @param writer  the output
 	 * @param matcher the matcher to use to split
-	 * @param max     the maximum number of mismatches
 	 * @param tags    the tags where the barcodes are
 	 */
-	private void runPaired(SamReader reader, SplitFastqWriter writer, BarcodeDecoder matcher, int[] max,
-		String[] tags, ProgressLoggerExtension progress) {
+	private void runPaired(SamReader reader, SplitFastqWriter writer, BarcodeDecoder matcher, String[] tags,
+		ProgressLoggerExtension progress) {
 		SAMRecordIterator it = reader.iterator();
 		while (it.hasNext()) {
 			SAMRecord record1 = it.next();
@@ -141,8 +145,8 @@ public class TaggedBamToFastq extends AbstractTool {
 			}
 			SAMRecord record2 = it.next();
 			String[] barcodes = getBarcodeFromTags(record1, tags);
-			String best = matcher.getBestBarcode(max, barcodes);
-			if (best.equals(BarcodeDecoder.UNKNOWN_STRING)) {
+			String best = matcher.getBestBarcode(barcodes);
+			if (best.equals(BarcodeMatch.UNKNOWN_STRING)) {
 				SAMRecordUtils.addBarcodeToName(record1, String.join("", barcodes));
 				SAMRecordUtils.addBarcodeToName(record2, String.join("", barcodes));
 			} else {
@@ -162,15 +166,14 @@ public class TaggedBamToFastq extends AbstractTool {
 	 * @param reader  the input reader
 	 * @param writer  the output
 	 * @param matcher the matcher to use to split
-	 * @param max     the maximum number of mismatches
 	 * @param tags    the tags where the barcodes are
 	 */
-	private void runSingle(SamReader reader, SplitFastqWriter writer, BarcodeDecoder matcher, int[] max,
-		String[] tags, ProgressLoggerExtension progress) {
+	private void runSingle(SamReader reader, SplitFastqWriter writer, BarcodeDecoder matcher, String[] tags,
+		ProgressLoggerExtension progress) {
 		for (SAMRecord record : reader) {
 			String[] barcodes = getBarcodeFromTags(record, tags);
-			String best = matcher.getBestBarcode(max, barcodes);
-			if (best.equals(BarcodeDecoder.UNKNOWN_STRING)) {
+			String best = matcher.getBestBarcode(barcodes);
+			if (best.equals(BarcodeMatch.UNKNOWN_STRING)) {
 				SAMRecordUtils.addBarcodeToName(record, String.join("", barcodes));
 			} else {
 				SAMRecordUtils.addBarcodeToName(record, best);
@@ -228,7 +231,7 @@ public class TaggedBamToFastq extends AbstractTool {
 								.hasArg().numberOfArgs(1).argName("BARCODES.tab").required().build();
 		Option max = Option.builder("m").longOpt("maximum-mismatches").desc(
 			"Maximum number of mismatches alowwed for a matched barcode. It could be provided only once for use in all barcodes or the same number of times as barcodes provided in the file. [Default="
-				+ BarcodeDecoder.DEFAULT_MISMATCHES + "]").hasArg().numberOfArgs(1).argName("INT")
+				+ BarcodeDecoder.DEFAULT_MAXIMUM_MISMATCHES + "]").hasArg().numberOfArgs(1).argName("INT")
 						   .required(false).build();
 		Option single = Option.builder("s").longOpt("single").desc("Switch to single-end parsing").hasArg(false)
 							  .required(false).build();