+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+### Changed
+- Changed logic of bacode detection: at least one difference between the best barcode and the second best is needed to do not discard a barcode
+### Added
+- Output metrics file when detecting barcodes
+
 ## [0.1.2] - 2015-11-26
-## Added
+### Added
 - Checking of output file existence
 - Create directories for output if they do not exists
 - Checking number of arguments